requirements.txt
fastapi~=0.105.0
starlette~=0.27.0
jinja2~=3.1.2
uvicorn~=0.24.0
pyotp~=2.9.0
fastapi-cache2~=0.2.1
